refactored config_file
